Import major rewrite of arachsys-a2
dnsmirror is reworked to support peers as well as slaves. We just rsync the
CDB file to the slaves as we needn't run tinydns-data repeatedly. dnsconfig
and dnssort are updated to reflect data file move from data to data.local.

We introduce new scripts mailuseradd, mailuserdel and mailpasswd, using
mailconfig in the intended race-free manner via open2().

mailconfig and dnsconfig now have different -s and non -s behaviour. With -s,
they die at the first error; without -s they check the whole input for errors
before dying. This is more useful than continuing after errors, which is very
unsafe behaviour.

The semantics on mailconfig for domains with multiple users are significantly
changed. The -a flag for non-privileged users is removed and replaced with a
new tool maildomains which lists domains with mail configuration belonging to
a given user, or all domains with corresponding users when run as root.